requirements/compiled.txt
# Choose your databases adapter
psycopg2

# for bcrypt passwords
# bcrypt

# Images
pillow
django-image-cropping==0.8.1
easy_thumbnails==1.5
